apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
PyJWT==2.10.1
cryptography==44.0.2
python-dotenv==1.1.0
openai>=1.30.0
